Move QueryFragment to using Display
